src/main/java/com/processout/payment/gateway/service/impl/MerchantService.java
package com.processout.payment.gateway.service.impl;

import com.processout.payment.gateway.model.Merchant;
import com.processout.payment.gateway.repository.MerchantRepository;
import com.processout.payment.gateway.service.IMerchantService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class MerchantService implements IMerchantService {

    private final MerchantRepository merchantRepository;

    @Autowired
    public MerchantService(MerchantRepository merchantRepository) {
        this.merchantRepository = merchantRepository;
    }

    // Example methods using the repository
    @Override
    public List<Merchant> getAllMerchants() {
        return merchantRepository.findAll();
    }

    @Override
    public Merchant getMerchantById(Long id) {
        return merchantRepository.findById(id).orElse(null);
    }

    @Override
    public Merchant saveMerchant(Merchant merchant) {
        return merchantRepository.save(merchant);
    }

    @Override
    public void deleteMerchant(Long id) {
        merchantRepository.deleteById(id);
    }
}
